fix: variable type

+++ inc/managers/sound_manager.hpp
@@ -26,16 +26,16 @@ class SoundManager {
   AdpcmResult playSfx(SoundFxCategory idCategory, SoundFX idSound,
                       const s8& t_ch);
 
-  void setSfxVolume(const s8& t_vol, const s8& t_ch);
+  void setSfxVolume(const u8& t_vol, const s8& t_ch);
 
-  const s8 getAvailableChannel();
+  const int getAvailableChannel();
 
  private:
   /**
    * @brief Declaration of global current adpcm channel
    *
    */
-  s8 currentAdpcmChannel = 0;
+  int currentAdpcmChannel = 0;
 
   SfxLibrary* soundLibrary;
   Engine* t_engine;
